Manage Listings, remove edit/delete from show page

+++ app/Http/Controllers/ListingController.php
@@ -47,6 +47,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message','Listing created successfully!');
@@ -63,6 +65,11 @@ class ListingController extends Controller
     public function update(Request $request, Listing $listing) {
         // dd($request->all());
         // dd($request->file('logo')->store());
+        //Make sure logged in user is ownwer
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+ 
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -84,7 +91,16 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        //Make sure logged in user is ownwer
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Manage Lisitngs
+    public function manage() {
+        return view('listings.manage',['listings' => auth()->user()->listings()->get()]);
     }
 }
